+++ LetsEncryptWinSimple.Core/Extensions/ObjectExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Reflection;
-using LetsEncryptWinSimple.Core.Extensions.ArrayExtensions;
 
 namespace LetsEncryptWinSimple.Core.Extensions
 {
@@ -92,52 +91,6 @@ namespace LetsEncryptWinSimple.Core.Extensions
         public override int GetHashCode(object obj)
         {
             return obj == null ? 0 : obj.GetHashCode();
-        }
-    }
-
-    namespace ArrayExtensions
-    {
-        public static class ArrayExtensions
-        {
-            public static void ForEach(this Array array, Action<Array, int[]> action)
-            {
-                if (array.LongLength == 0) return;
-                var walker = new ArrayTraverse(array);
-
-                do action(array, walker.Position);
-                while (walker.Step());
-            }
-        }
-
-        internal class ArrayTraverse
-        {
-            public int[] Position;
-            private readonly int[] _maxLengths;
-
-            public ArrayTraverse(Array array)
-            {
-                _maxLengths = new int[array.Rank];
-                for (var i = 0; i < array.Rank; ++i)
-                    _maxLengths[i] = array.GetLength(i) - 1;
-
-                Position = new int[array.Rank];
-            }
-
-            public bool Step()
-            {
-                for (var i = 0; i < Position.Length; ++i)
-                {
-                    if (Position[i] >= _maxLengths[i])
-                        continue;
-
-                    Position[i]++;
-                    for (var j = 0; j < i; j++)
-                        Position[j] = 0;
-
-                    return true;
-                }
-                return false;
-            }
         }
     }
 }
